bring to a working state
