+++ src/main/java/org/terasology/las/platform/FloatingPlatformProvider.java
@@ -16,7 +16,6 @@
 
 package org.terasology.las.platform;
 
-import org.joml.Vector3ic;
 import org.terasology.engine.entitySystem.Component;
 import org.terasology.engine.world.block.BlockArea;
 import org.terasology.engine.world.block.BlockAreac;
@@ -24,12 +23,9 @@ import org.terasology.engine.world.block.BlockRegion;
 import org.terasology.engine.world.block.BlockRegionc;
 import org.terasology.engine.world.generation.Border3D;
 import org.terasology.engine.world.generation.ConfigurableFacetProvider;
-import org.terasology.engine.world.generation.Facet;
 import org.terasology.engine.world.generation.FacetProviderPlugin;
 import org.terasology.engine.world.generation.GeneratingRegion;
 import org.terasology.engine.world.generation.Produces;
-import org.terasology.engine.world.generation.Updates;
-import org.terasology.engine.world.generation.facets.SurfacesFacet;
 import org.terasology.engine.world.generator.plugin.RegisterPlugin;
 import org.terasology.ligthandshadow.componentsystem.LASUtils;
 
@@ -41,7 +37,6 @@ import java.util.Collections;
  */
 @RegisterPlugin
 @Produces(FloatingPlatformFacet.class)
-@Updates(@Facet(value = SurfacesFacet.class))
 public class FloatingPlatformProvider implements ConfigurableFacetProvider, FacetProviderPlugin {
     private static final BlockAreac FLOATING_PLATFORM_REGION =
             new BlockArea(LASUtils.FLOATING_PLATFORM_POSITION.x() - LASUtils.FLOATING_PLATFORM_WIDTH / 2,
@@ -72,7 +67,6 @@ public class FloatingPlatformProvider implements ConfigurableFacetProvider, Face
 
     @Override
     public void process(GeneratingRegion region) {
-        SurfacesFacet surfacesFacet = region.getRegionFacet(SurfacesFacet.class);
         Border3D border = region.getBorderForFacet(FloatingPlatformFacet.class);
         FloatingPlatformFacet platformFacet = new FloatingPlatformFacet(region.getRegion(), border);
         BlockAreac worldRect = platformFacet.getWorldArea();
@@ -83,15 +77,7 @@ public class FloatingPlatformProvider implements ConfigurableFacetProvider, Face
                 platformFacet.add(platform);
             }
         }
-        BlockRegion surfaceRect = surfacesFacet.getWorldRegion();
-        for (Vector3ic pos : surfaceRect) {
-            if (FLOATING_PLATFORM_REGION.contains(pos.x(), pos.z())) {
-                int y = surfacesFacet.getNextBelow(pos);
-                surfacesFacet.setWorld(pos.x(), y, pos.z(), false);
-            }
-        }
         region.setRegionFacet(FloatingPlatformFacet.class, platformFacet);
-        region.setRegionFacet(SurfacesFacet.class, surfacesFacet);
     }
 
     @Override
